requirements.txt
google-generativeai
requests
python-dotenv
edge-tts
moviepy
pillow
openai-whisper